+++ ubo_app/store/services/notifications.py
@@ -89,7 +89,7 @@ class Notification(Immutable):
     content: str
     extra_information: str | None = None
     importance: Importance = Importance.LOW
-    chime: Chime = Chime.DONE
+    chime: Chime | None = Chime.DONE
     timestamp: datetime = field(default_factory=lambda: datetime.now(tz=UTC))
     is_read: bool = False
     sender: str | None = None
@@ -102,6 +102,8 @@ class Notification(Immutable):
     dismissable: bool = True
     dismiss_on_close: bool = False
     on_close: Callable[[], Any] | None = None
+    progress: float | None = None
+    progress_weight: float = 1
 
 
 class NotificationsAction(BaseAction): ...
@@ -134,3 +136,4 @@ class NotificationsDisplayEvent(NotificationsEvent):
 class NotificationsState(Immutable):
     notifications: Sequence[Notification]
     unread_count: int
+    progress: float | None = None