+++ ubo_app/utils/async_.py
@@ -27,8 +27,6 @@ def create_task(
 
         logger = get_logger('ubo-app')
 
-        if awaitable is None:
-            return
         try:
             thread = current_thread()
             logger.verbose(