+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - fix(core): add `dtoverlay=spi0-0cs` to `/boot/firmware/config.txt` to make the LCD work on RPi5
 - refactor: general housekeeping, improving typing, linting, resource management, etc
 - fix(notifications): avoid auto-closing notifications shown in the notification center
+- feat(camera): fail-proof the camera initialization when no camera is connected
 
 ## Version 0.15.3